+++ crates/azure/src/compute_publisher_vm_image_offer_skus.rs
@@ -78,12 +78,12 @@ cloud_terrastodon_command::impl_cacheable_into_future!(ComputePublisherImageOffe
 mod test {
     use crate::prelude::fetch_all_subscriptions;
     use crate::prelude::fetch_compute_publisher_image_offer_skus;
-    use crate::prelude::get_default_tenant_id;
+    use crate::prelude::get_test_tenant_id;
     use cloud_terrastodon_azure_types::prelude::LocationName;
 
     #[tokio::test]
     pub async fn it_works() -> eyre::Result<()> {
-        let tenant_id = get_default_tenant_id().await?;
+        let tenant_id = get_test_tenant_id().await?;
         let subscription_id = fetch_all_subscriptions(tenant_id)
             .await?
             .first()