+++ crates/azure/src/virtual_machine_sizes.rs
@@ -34,12 +34,12 @@ pub async fn fetch_virtual_machine_sizes(
 #[cfg(test)]
 mod test {
     use crate::prelude::fetch_all_subscriptions;
-    use crate::prelude::get_default_tenant_id;
+    use crate::prelude::get_test_tenant_id;
     use cloud_terrastodon_azure_types::prelude::LocationName;
 
     #[tokio::test]
     pub async fn it_works() -> eyre::Result<()> {
-        let tenant_id = get_default_tenant_id().await?;
+        let tenant_id = get_test_tenant_id().await?;
         let subs = fetch_all_subscriptions(tenant_id).await?;
         let sub = subs.first().unwrap();
         let sizes =